Add separate MarkovKernel class
