+++ nfmc/algorithms/mh/base.py
@@ -1,8 +1,10 @@
 from typing import Tuple, Union
 import torch
 
+from nfmc.algorithms.kernel import MarkovKernel
 
-class MHKernel:
+
+class MHKernel(MarkovKernel):
     """
     Base MCMC kernel class for Metropolis-Hastings algorithms.
     """
@@ -18,20 +20,10 @@ class MHKernel:
          It receives as input a tensor with shape `(*batch_shape, *event_shape)` and outputs a tensor with shape
          `batch_shape`.
         """
-        self.event_shape = event_shape
-        self.neg_log_prob_target = neg_log_prob_target
+        super().__init__(event_shape, neg_log_prob_target)
 
-        self._n_steps: int = 0
         self._n_attempted_transitions: int = 0
         self._n_accepted_transitions: int = 0
-        self._n_calls: int = 0  # Target density evaluation counter
-        self._n_grads: int = 0  # Target density gradient evaluation counter
-        # Counts the number of chains that diverged across all steps
-        self._n_divergences: int = 0
-
-    @property
-    def name(self) -> str:
-        raise NotImplementedError
 
     def reset_statistics(self):
         self._n_steps = 0
@@ -47,9 +39,6 @@ class MHKernel:
             return torch.nan
         return self._n_accepted_transitions / self._n_attempted_transitions
 
-    def increment_n_steps(self):
-        self._n_steps += 1
-
     def increment_n_attempted_transitions(self, n_chains: int):
         self._n_attempted_transitions += n_chains
         self._n_attempted_transitions = int(self._n_attempted_transitions)
@@ -58,35 +47,4 @@ class MHKernel:
         self._n_accepted_transitions += n_accepted_chains
         self._n_accepted_transitions = int(self._n_accepted_transitions)
 
-    def increment_n_calls(self, n_calls: int):
-        self._n_calls += n_calls
-        self._n_calls = int(self._n_calls)
 
-    def increment_n_grads(self, n_grads: int):
-        self._n_grads += n_grads
-        self._n_grads = int(self._n_grads)
-
-    def increment_n_divergences(self, n_divergences: int):
-        self._n_divergences += n_divergences
-        self._n_divergences = int(self._n_divergences)
-
-    @property
-    def event_size(self):
-        return int(torch.prod(torch.as_tensor(self.event_shape)))
-
-    def step(self,
-             x: torch.Tensor,
-             *args,
-             update: bool = False,
-             **kwargs) -> Tuple[torch.Tensor, torch.Tensor]:
-        """
-        Performs one kernel transition.
-
-        :param torch.Tensor x: current state tensor with shape `(*batch_shape, *event_shape)`.
-        :param bool update: if True, update kernel parameters.
-        :return: new state tensor with shape `(*batch_shape, *event_shape)`.
-        """
-        raise NotImplementedError
-
-    def __repr__(self):
-        raise NotImplementedError
